refactor(gh): share repository identity validation
