+++ wiki/testing.md
@@ -72,7 +72,7 @@ task default: :test
 | `dependencies_test.rb`, `task_meta_test.rb`, `plan_frontmatter_test.rb`, `repository_identity_test.rb`, `dependency_admission_test.rb`, `dependency_snapshot_test.rb` | Fail-closed dependency admission — exhaustive scalar grammar and duplicate-key rejection; tolerant-vs-strict metadata reads and corrupt-mutation refusal; bounded optional exact plan assertion; SSH/HTTPS/local remote normalization plus bounded TERM/KILL lookup; indexed same/cross-project and archived-fallback resolution; full cycle paths; concurrent folder-move fencing; workflow/gate validation; clear/wait/error verdicts; and unexpected-error backstop. |
 | `integration/dependency_admission_test.rb`, `run_approve_test.rb`, `run_stage_action_test.rb`, `new_test.rb` | Supported-boundary coverage — anonymized plan-only ordering and repository-mismatch fixtures, fresh validation before run/rebase/runner or forward approve mutation, retryable waits, non-retryable admission errors, `--force` refusal, backward recovery, composed JSON propagation, and all three creation forms. |
 | `git_ops_test.rb` | `Hive::GitOps` — default-branch detection, orphan worktree bootstrap, idempotent gitignore, empty-diff commit skip. |
-| `gh_test.rb` | `Hive::Gh` — PR frontmatter parsing, secret-scan fetch-failure semantics, open/merged PR lookup, `pr_state` success/error parsing, `pr_metadata` parsing/error handling/project `chdir:` scoping, `PushResult`, process-group timeout/termination including a TERM-resistant stdout-inheriting helper, exact created-PR head/base OID identity checks, `mergeStateStatus` request shape for open PR listing, merged-PR digest repo discovery/listing helpers, and failing-job log clipping. |
+| `gh_test.rb` | `Hive::Gh` — PR frontmatter parsing, secret-scan fetch-failure semantics, open/merged PR lookup, `pr_state` success/error parsing, `pr_metadata` parsing/error handling/project `chdir:` scoping, `PushResult`, process-group timeout/termination including a TERM-resistant stdout-inheriting helper, shared strict GitHub host/repository validation across normal and architecture-patrol transports, exact created-PR head/base OID identity checks, `mergeStateStatus` request shape for open PR listing, merged-PR digest repo discovery/listing helpers, and failing-job log clipping. |
 | `pr_test.rb` | `Hive::Pr` — pull-request-number extraction from `/pull/<number>` URLs, including query/fragment/trailing-slash tolerance, nil for issue/non-number/subpage URLs, `identifier_to_number` acceptance/rejection for `hive review --pr`, and http(s) URL validation including invalid-URI rejection. |
 | `agent_limit_test.rb` | `Hive::AgentLimit` — provider-limit classifier for Claude usage-credit menus and common quota/rate-limit API errors, retry-cooldown timestamp helpers, and the shared quota-held display/JSON helpers (`held?`, provider extraction, UTC retry display, label text, and `held` field shape), with runnable false-positive guards for source line numbers and ordinary "missing rate limit" findings. The current file also contains intended UI-feature limit assertions, but they sit below a `private` marker; this refresh treats them as a coverage gap until moved into the public runnable test set. |
 | `agent_test.rb`, `agent_profile_test.rb` | `Hive::Agent` / `AgentProfile` — spawn/wait/timeout/SIGINT forwarding, version and capability checks, configured permission/workspace-write argv, Claude model/effort and verified safe-mode `cli_flags`, process-group cleanup for timed-out version/help probes and stdout-inheriting descendants, safe-mode non-leakage to unrelated launches, and provider-limit classification before generic exit-code / expected-output failures. |
